refactor(ingest): decompose ast_walker.go into focused files [mache-godfile-ast-walker]
ast_walker.go was 1080 LOC mixing six concerns: walker core, selector
parsing, SQL node lookups, call-pattern extraction, value/context/imports
extraction, and the Match implementation. Decomposed into seam-aligned
files; the public API (ASTWalker methods, ASTRoot, CallPattern,
Register* helpers, SelectWalker) is unchanged.

Files:
- ast_walker.go (241): ASTWalker, NewASTWalker, EnsureIndexes, ASTRoot,
  Query, Close, SelectWalker, astNode, readSource
- ast_walker_selector.go (356): selectorPattern + parseSelector +
  S-expr tokenizer + ancestry matching/stripping
- ast_walker_nodes.go (124): findNodesByKind, findChildByKindAST
- ast_walker_calls.go (203): CallPattern + registry + ExtractCalls +
  ExtractQualifiedCalls + queryCallPattern
- ast_walker_extract.go (163): ExtractAddressRefs, contextKindRegistry,
  ExtractContext, ExtractGoImports
- ast_walker_match.go (28): astMatch + CaptureOrigin

New tests (ast_walker_core_test.go, 8 tests) cover pre-existing branches
the coverage gate flagged after the move (rename detection couldn't
follow at -B5% -M5% -C5% because ast_walker.go is only 22% similar to
its predecessor): EnsureIndexes idempotency, Close no-op, $ wildcard
selector, wrong-root-type rejection, missing required vs. underscored
captures, byte-range fallback when leaf records are empty, readSource
disk-path fallback and no-content+no-path error.

One unreachable defensive branch (`if cap.name == "scope"` inside the
capture loop) is annotated `// coverage:ignore` — parseSelector strips
@scope captures before they reach pattern.captures, so the guard cannot
fire from the public API; kept for future selector-grammar evolution.

Validation:
- coverage-gate: 0 new uncovered lines (at -rename-threshold 50)
- task check: exit 0
- go test -race -count=2 ./internal/ingest/...: PASS
- TestASTParity_{Go,Python,Elixir} (build tag leyline): PASS
